+++ ingresos_egresos/src/main/java/servicios/exequiales/ingresosyegresos/ingresos_egresos/Service/IEmpleadoService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface IEmpleadoService {
 
-    public Empleado findById(int id);
+    public Empleado findById(long id);
     public List<Empleado> findAll();
     public Empleado createEmpleado(Empleado empleado);
     public Empleado updateEmpleado(int id, Empleado empleado);
